Introduced config framework + made bitcoin.conf location configurable
The config framework allows the RPC client to be configured via a
config.properties file. The config values are read programatically and
can have any data type (the framework converts and casts automatically
to the right type).

As an example, introduced a config paramter for the location of the
bitcoin core data folder. This allows the caller to specify a custom
location for this folder, where the RPC client will look for the
bitcoin.conf file.